microservice-db/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
marshmallow==3.21.1
bcrypt==4.1.2 